+++ src/steps/constants.ts
@@ -1,6 +1,8 @@
 import {
   RelationshipClass,
+  RelationshipDirection,
   StepEntityMetadata,
+  StepMappedRelationshipMetadata,
   StepRelationshipMetadata,
 } from '@jupiterone/integration-sdk-core';
 
@@ -9,12 +11,14 @@ export const Steps = {
   USERS: 'fetch-users',
   GROUPS: 'fetch-groups',
   ROLES: 'fetch-roles',
+  ASSETS: 'fetch-assets',
+  FINDINGS: 'fetch-findings',
   GROUP_USER_RELATIONSHIPS: 'build-user-group-relationships',
   USER_ROLE_RELATIONSHIPS: 'build-user-role-relationships',
 };
 
 export const Entities: Record<
-  'ACCOUNT' | 'GROUP' | 'USER' | 'ROLE',
+  'ACCOUNT' | 'GROUP' | 'USER' | 'ROLE' | 'ASSET' | 'FINDING' | 'CVE',
   StepEntityMetadata
 > = {
   ACCOUNT: {
@@ -64,13 +68,66 @@ export const Entities: Record<
       required: ['name'],
     },
   },
+  ASSET: {
+    resourceName: 'Asset',
+    _type: 'orca_asset',
+    _class: ['Resource'],
+    schema: {
+      properties: {
+        name: { type: 'string' },
+      },
+      required: ['name'],
+    },
+  },
+  FINDING: {
+    resourceName: 'Finding',
+    _type: 'orca_finding',
+    _class: ['Finding'],
+    schema: {
+      properties: {
+        name: { type: 'string' },
+      },
+      required: ['name'],
+    },
+  },
+  CVE: {
+    resourceName: 'CVE',
+    _type: 'cve',
+    _class: ['Vulnerability'],
+    schema: {
+      properties: {
+        name: { type: 'string' },
+        displayName: { type: 'string' },
+        cvssScore: { type: 'string' },
+        references: { type: 'array', items: { type: 'string' } },
+        webLink: { type: 'string' },
+      },
+      required: ['name', 'displayName', 'cvssScore', 'references', 'weblink'],
+    },
+  },
+};
+
+export const MappedRelationships: Record<
+  'FINDING_IS_CVE',
+  StepMappedRelationshipMetadata
+> = {
+  FINDING_IS_CVE: {
+    _type: 'orca_finding_is_cve',
+    sourceType: Entities.FINDING._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.CVE._type,
+    direction: RelationshipDirection.FORWARD,
+  },
 };
 
 export const Relationships: Record<
   | 'ACCOUNT_HAS_USER'
   | 'ACCOUNT_HAS_GROUP'
+  | 'ACCOUNT_HAS_ASSET'
   | 'GROUP_HAS_USER'
-  | 'USER_ASSIGNED_ROLE',
+  | 'USER_ASSIGNED_ROLE'
+  | 'ACCOUNT_HAS_FINDING'
+  | 'ASSET_HAS_FINDING',
   StepRelationshipMetadata
 > = {
   ACCOUNT_HAS_USER: {
@@ -85,6 +142,12 @@ export const Relationships: Record<
     _class: RelationshipClass.HAS,
     targetType: Entities.GROUP._type,
   },
+  ACCOUNT_HAS_ASSET: {
+    _type: 'orca_account_has_asset',
+    sourceType: Entities.ACCOUNT._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.ASSET._type,
+  },
   GROUP_HAS_USER: {
     _type: 'orca_group_has_user',
     sourceType: Entities.GROUP._type,
@@ -96,5 +159,17 @@ export const Relationships: Record<
     sourceType: Entities.USER._type,
     _class: RelationshipClass.ASSIGNED,
     targetType: Entities.ROLE._type,
+  },
+  ACCOUNT_HAS_FINDING: {
+    _type: 'orca_account_has_finding',
+    sourceType: Entities.ACCOUNT._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.FINDING._type,
+  },
+  ASSET_HAS_FINDING: {
+    _type: 'orca_asset_has_finding',
+    sourceType: Entities.ASSET._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.FINDING._type,
   },
 };